Simplify news and services
We're now only retrieving 3 items per section so we can delete a lot of code
from the news_and_communications and services presenters and get them to return
an array of results rather than hashes.

This allows us to get them to behave a bit more like the other presenters.